requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
motor==3.6.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.6.0
pymongo==4.9.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.20.0
